+++ railties/CHANGELOG.md
@@ -1,3 +1,3 @@
-
+* You can now refer to `__FILE__` when parsing ERB in `config/database.yml`
 
 Please check [6-0-stable](https://github.com/rails/rails/blob/6-0-stable/railties/CHANGELOG.md) for previous changes.